Modified to use Spring MVC, Hibernate

+++ scripts/create.sql
@@ -1,6 +1,6 @@
 CREATE TABLE user
 (
-    userid   INT         NOT NULL AUTO_INCREMENT,
+    userid   BIGINT      NOT NULL AUTO_INCREMENT,
     name     VARCHAR(45) NOT NULL,
     sureName VARCHAR(45) NOT NULL,
 
@@ -9,9 +9,9 @@ CREATE TABLE user
 
 CREATE TABLE account
 (
-    accountid INT NOT NULL AUTO_INCREMENT,
-    account   INT NOT NULL,
-    userid    INT NOT NULL,
+    accountid BIGINT NOT NULL AUTO_INCREMENT,
+    account   BIGINT NOT NULL,
+    userid    BIGINT NOT NULL,
 
     PRIMARY KEY (accountid),
     FOREIGN KEY (userid)
